requirements.txt
Pillow>=9.2.0
deepl>=1.15.0 